backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
pymongo==4.6.1
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0
python-dotenv==1.0.0

